# JDK-8314056 Remove runtime platform check from frem/drem