Remove platform check and move code to stubGenerator.  This fix increases performance by ~4.5%.

UPDATE: Subsequent commits increase performance gain to ~2x for AVX2, with no significant change to AVX512.

Tested tier1.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8314056](https://bugs.openjdk.org/browse/JDK-8314056): Remove runtime platform check from frem/drem (**Enhancement** - P4)


### Reviewers
 * [Sandhya Viswanathan](https://openjdk.org/census#sviswanathan) (@sviswa7 - **Reviewer**)
 * [Jatin Bhateja](https://openjdk.org/census#jbhateja) (@jatin-bhateja - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/15210/head:pull/15210` \
`$ git checkout pull/15210`

Update a local copy of the PR: \
`$ git checkout pull/15210` \
`$ git pull https://git.openjdk.org/jdk.git pull/15210/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 15210`

View PR using the GUI difftool: \
`$ git pr show -t 15210`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/15210.diff">https://git.openjdk.org/jdk/pull/15210.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/15210#issuecomment-1671831668)