@sviswa7 2023-08-24
The changes look good to me now. Could you please also get the performance of DremFrem jmh micro with/without your PR for both UseAVX=2 and UseAVX=3?