@asgibbons 2023-08-27
> The changes look good to me now. Could you please also get the performance of DremFrem jmh micro with/without your PR for both UseAVX=2 and UseAVX=3?

Results of DremFrem:

Before these changes:
```
java -XX:UseAVX=2 -jar ./build/linux-x86_64-server-release/images/test/micro/benchmarks.jar -f 1 DremFrem:

Benchmark                      Mode  Cnt  Score   Error  Units
DremFrem.calcDoubleJava        avgt    5  5.645 ± 0.003  ns/op
DremFrem.calcFloatJava         avgt    5  6.044 ± 0.005  ns/op
DremFrem.cornercaseDoubleJava  avgt    5  5.462 ± 0.033  ns/op
DremFrem.cornercaseFloatJava   avgt    5  5.494 ± 0.045  ns/op

java -XX:UseAVX=3 -jar ./build/linux-x86_64-server-release/images/test/micro/benchmarks.jar -f 1 DremFrem:

Benchmark                      Mode  Cnt  Score   Error  Units
DremFrem.calcDoubleJava        avgt    5  2.916 ± 0.006  ns/op
DremFrem.calcFloatJava         avgt    5  4.016 ± 0.001  ns/op
DremFrem.cornercaseDoubleJava  avgt    5  5.506 ± 0.009  ns/op
DremFrem.cornercaseFloatJava   avgt    5  5.506 ± 0.010  ns/op

```
=============================================
WITH these changes:

```
java -XX:UseAVX=2 -jar ./build/linux-x86_64-server-release/images/test/micro/benchmarks.jar -f 1 DremFrem:

Benchmark                      Mode  Cnt  Score   Error  Units
DremFrem.calcDoubleJava        avgt    5  2.838 ± 0.001  ns/op
DremFrem.calcFloatJava         avgt    5  3.554 ± 0.002  ns/op
DremFrem.cornercaseDoubleJava  avgt    5  5.511 ± 0.022  ns/op
DremFrem.cornercaseFloatJava   avgt    5  5.467 ± 0.014  ns/op

java -XX:UseAVX=3 -jar ./build/linux-x86_64-server-release/images/test/micro/benchmarks.jar -f 1 DremFrem:

Benchmark                      Mode  Cnt  Score    Error  Units
DremFrem.calcDoubleJava        avgt    5  2.598 ±  0.001  ns/op
DremFrem.calcFloatJava         avgt    5  3.229 ±  0.001  ns/op
DremFrem.cornercaseDoubleJava  avgt    5  5.508 ±  0.014  ns/op
DremFrem.cornercaseFloatJava   avgt    5  5.511 ±  0.005  ns/op

```